세팅 버그 수정 및 메인 대시보드 작업 중
