+++ client/src/styles/intro.scss
@@ -24,7 +24,7 @@ $blue: #61759f;
     img {
       width: 300px;
       height: 300px;
-      margin-bottom: 10px;
+      margin-bottom: 60px;
     }
   }
 
@@ -36,6 +36,7 @@ $blue: #61759f;
     transition: visibility 0s, opacity 1s ease;
     width: 100%;
     height: 15%;
+    margin-top: 10px;
   }
 
   .pc-background,
