+++ client/src/styles/setting.scss
@@ -177,7 +177,7 @@ $blue: #61759f;
 
   .studentInfo-upload {
     margin: 10px 0 0 5px;
-    width: 20%;
+    width: 35%;
     padding: 4px 6px;
     font-size: 0.77rem;
     border: 1px solid #a0a895;
